requirements.txt
pip==24.0
setuptools==58.2.0
wheel
reportlab
python-docx

# Core Libraries
requests
matplotlib
urllib3
pandas
numpy
scipy
sqlalchemy
fastapi
uvicorn
watchdog
nest-asyncio
python-dotenv
nltk
fuzzywuzzy
python-Levenshtein
ngrok

# NLP & AI
spacy==3.8.4
thinc==8.3.4
sentence-transformers==3.3.1
transformers==4.48.1
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-experimental==0.3.4
langchain-text-splitters==0.3.5
langchain-groq==0.2.3
replicate==1.0.4
huggingface-hub==0.27.1

# Data Processing
pandasql==0.7.3
scikit-learn==1.6.1
pyarrow==19.0.0
openpyxl==3.1.5
jsonschema==4.23.0
jsonpatch==1.33
pydantic==2.10.5
pydantic-settings==2.7.1

# Web & APIs
httpx==0.28.1
httpx-sse==0.4.0
starlette==0.45.3
tenacity==9.0.0

# Streamlit & UI
streamlit==1.41.1
altair==5.5.0
pydeck==0.9.1
rich==13.9.4

# Misc
typer==0.15.1
tqdm==4.67.1
jinja2==3.1.5
pillow==11.1.0
filelock==3.16.1
sympy==1.13.1
yarl==1.18.3
blinker==1.9.0